AppDelegate // Make UpdateAPI into a standalone module.
